requirements.txt
pandas==1.2.3
numpy==1.19.5
matplotlib==3.2.1
opencv_python==4.3.0.36
boto3==1.17.49
Pillow==8.2.0
scikit_learn==0.24.1
openpyxl==3.0.7
tqdm==4.48.0
xlsxwriter==1.3.8
argparse==1.1
